+++ include/CDLListNode.h
@@ -5,14 +5,14 @@
 struct CDLListNode
 {
     // Read-write.
-    int item;
+    int long item;
 
     // Read-only.
     struct CDLListNode *prev;
     struct CDLListNode *next;
 };
 
-struct CDLListNode *CDLListNode_new(int item);
+struct CDLListNode *CDLListNode_new(int long item);
 void CDLListNode_delete(struct CDLListNode* node);
 
 #endif  // CIRCULAR_DOUBLY_LINKED_LIST_JOSEPHUS_PROBLEM_INCLUDE_CDLLISTNODE_H_
